Migrate to supporting passing in a webhook which can run as an arg in a default func
