+++ source/github/events/github_runner.ts
@@ -1,24 +1,18 @@
 import * as express from "express"
-import winston from "../../logger"
-
-import { PERIL_BOT_USER_ID } from "../../globals"
 
 import { DangerResults } from "danger/distribution/dsl/DangerResults"
-import { GitHub } from "danger/distribution/platforms/GitHub"
-import { GitHubAPI } from "danger/distribution/platforms/github/GitHubAPI"
-
-import vm2 from "danger/distribution/runner/runners/vm2"
 import { getTemporaryAccessTokenForInstallation } from "../../api/github"
 import { DangerRun, dangerRunForRules, dsl, feedback } from "../../danger/danger_run"
-import { executorForInstallation, runDangerForInstallation } from "../../danger/danger_runner"
-import perilPlatform from "../../danger/peril_platform"
 import { GitHubInstallation, GithubRepo } from "../../db"
 import { getDB } from "../../db/getDB"
 import { GitHubInstallationSettings } from "../../db/GitHubRepoSettings"
 import logger from "../../logger"
-import { Pull_request } from "../events/types/pull_request_opened.types"
-import { canUserWriteToRepo, getGitHubFileContents } from "../lib/github_helpers"
-import { createPRDSL } from "./createPRDSL"
+import winston from "../../logger"
+import { runEventRun } from "./handlers/event"
+import { runPRRun } from "./handlers/pr"
+import { actionForWebhook } from "./utils/actions"
+import { commentOnResults, getIssueNumber } from "./utils/commenting"
+import { repoIsIgnored } from "./utils/ignore_repos"
 
 /**
  * So, these function have a bunch of responsibilities.
@@ -93,27 +87,6 @@ export const setupForRequest = async (req: express.Request, installationSettings
   }
 }
 
-/**
- * There's value in not running Dangerfiles for every webhook
- * mainly that it's wasteful, so this gives us
- */
-export const actionForWebhook = (webhook: any): string | null => {
-  // PR/Issues etc
-  if (webhook.action) {
-    return webhook.action
-  }
-
-  // Are there more worth adding?
-
-  // Fallback for Status
-  // https://developer.github.com/v3/activity/events/types/#statusevent
-  if (webhook.state) {
-    return webhook.state
-  }
-
-  return null
-}
-
 export const githubDangerRunner = async (event: string, req: express.Request, res: express.Response, next: any) => {
   const action = actionForWebhook(req.body)
   const installationID = req.body.installation.id as number
@@ -135,16 +108,11 @@ export const githubDangerRunner = async (event: string, req: express.Request, re
 
   // Allow edge-case repos to skip Danger rules. E.g. in Artsy, our analytics and marketing repos
   // do not need the same level of thought as an larger engineering project would.
-  if (
-    settings.repoName &&
-    installation.settings &&
-    installation.settings.ignored_repos &&
-    installation.settings.ignored_repos.includes(settings.repoName)
-  ) {
+  if (repoIsIgnored(settings.repoName, installation)) {
     res.status(204).send({ message: `Skipping peril run due to repo being in ignored`, iID: installationID })
     return
   }
-  winston.info("Got past")
+
   // Some events aren't tied to a repo (like creating a user) and so
   // right now I've not thought through what is necessary to run those
   if (!settings.isRepoEvent) {
@@ -192,7 +160,6 @@ export const runEverything = async (
   }
 
   const token = await getTemporaryAccessTokenForInstallation(req.body.installation.id)
-
   const allResults = [] as DangerResults[]
 
   const prRuns = runs.filter(r => r.dslType === dsl.pr)
@@ -226,158 +193,6 @@ export const runEverything = async (
   res.status(200).send(JSON.stringify({ status, results: allResults }, null, "  "))
 }
 
-export const runEventRun = async (
-  run: DangerRun,
-  settings: GitHubRunSettings,
-  token: string,
-  dangerDSL: any
-): Promise<DangerResults | null> => {
-  const repoForDangerfile = run.repoSlug || (dangerDSL.repository && dangerDSL.repository.full_name)
-  if (!repoForDangerfile) {
-    return null
-  }
-
-  const supportsGithubCommentAPIs = run.feedback === feedback.commentable
-
-  // Do we need an authenticated Danger GitHubAPI instance so we
-  // can leave feedback on an issue?
-  let githubAPI = null as GitHubAPI | null
-  if (supportsGithubCommentAPIs && settings.commentableID && settings.repoName) {
-    githubAPI = githubAPIForCommentable(token, settings.repoName, settings.commentableID)
-  }
-
-  const headDangerfile = await getGitHubFileContents(token, repoForDangerfile, run.dangerfilePath, run.branch)
-
-  const installationSettings = {
-    iID: settings.installationID,
-    settings: settings.installationSettings,
-  }
-
-  const results = await runDangerForInstallation(
-    headDangerfile,
-    run.referenceString,
-    githubAPI,
-    run.dslType,
-    installationSettings,
-    { github: dangerDSL }
-  )
-
-  return results || null
-}
-
-export const runPRRun = async (
-  run: DangerRun,
-  settings: GitHubRunSettings,
-  token: string,
-  pr: Pull_request
-): Promise<DangerResults | null> => {
-  if (!settings.repoName) {
-    console.error("An event without a repo name was passed to runRPRun") // tslint:disable-line
-    return null
-  }
-
-  if (!settings.triggeredByUsername) {
-    console.error("An event without a username was passed to runRPRun") // tslint:disable-line
-    return null
-  }
-
-  if (!pr.head.repo) {
-    console.error("An event without a head repo was passed to runPRRun") // tslint:disable-line
-    return null
-  }
-
-  const githubAPI = githubAPIForCommentable(token, settings.repoName, settings.commentableID)
-
-  // In theory only a PR requires a custom branch, so we can check directly for that
-  // in the event JSON and if it's not there then use master
-  // prioritise the run metadata
-
-  const dangerfileRepoForPR = pr.head.repo.full_name
-  const dangerfileBranchForPR = pr.head.ref
-  const neededDangerfileIsLocalRepo = !run.repoSlug
-  const branch = neededDangerfileIsLocalRepo ? dangerfileBranchForPR : null
-
-  // Either it's dictated in the run as an external repo, or we use the most natural repo
-  const repoForDangerfile = run.repoSlug || dangerfileRepoForPR
-
-  const baseDangerfile = await getGitHubFileContents(token, repoForDangerfile, run.dangerfilePath, branch)
-  const headDangerfile = await getGitHubFileContents(token, repoForDangerfile, run.dangerfilePath, branch)
-  const dangerfilesExist = headDangerfile !== "" && baseDangerfile !== ""
-
-  // Shortcut to determine if both Dangerfile exists, and that they have different content
-  if (dangerfilesExist && baseDangerfile !== headDangerfile) {
-    // Check to see if they have write access, if they don't then don't run the
-    // Dangerfile, but put out a message that it's not being ran on purpose
-    const userCanWrite = await canUserWriteToRepo(token, settings.triggeredByUsername, dangerfileRepoForPR)
-    if (!userCanWrite) {
-      const message = "Not running Danger rules due to user with no write access changing the Dangerfile."
-      return {
-        fails: [],
-        markdowns: [],
-        warnings: [],
-        messages: [{ message }],
-      }
-    }
-  }
-
-  const reportData = (reason: string) => {
-    const stateForErrorHandling = {
-      branch,
-      dangerfileBranchForPR,
-      neededDangerfileIsLocalRepo,
-      repoForDangerfile,
-      run,
-      settings,
-    }
-
-    return `${reason}
-
-## Full state of PR run:
-
-\`\`\`json
-${JSON.stringify(stateForErrorHandling, null, "  ")}
-\`\`\`
-      `
-  }
-
-  if (headDangerfile === "") {
-    const actualBranch = branch ? branch : "master"
-    const message = `Could not find Dangerfile at <code>${
-      run.dangerfilePath
-    }</code> on <code>${repoForDangerfile}</code> on branch <code>${actualBranch}</code>`
-
-    const report = reportData(message)
-    return {
-      fails: [{ message: report }],
-      markdowns: [],
-      warnings: [],
-      messages: [],
-    }
-  } else {
-    // Everything is :+1:
-
-    const installationSettings = {
-      iID: settings.installationID,
-      settings: settings.installationSettings,
-    }
-
-    const dangerDSL = await createPRDSL(githubAPI)
-    const results = await runDangerForInstallation(
-      headDangerfile,
-      run.referenceString,
-      githubAPI,
-      run.dslType,
-      installationSettings,
-      dangerDSL
-    )
-
-    if (results && pr.body && pr.body.includes("Peril: Debug")) {
-      results.markdowns.push({ message: reportData("Showing PR details due to including 'Peril: Debug'") })
-    }
-    return results ? results : null
-  }
-}
-
 export const mdResults = (results: DangerResults): string => {
   return `
 mds: ${results.markdowns.length}
@@ -399,46 +214,6 @@ export const mergeResults = (results: DangerResults[]): DangerResults => {
     },
     { fails: [], markdowns: [], warnings: [], messages: [] }
   )
-}
-
-export const commentOnResults = async (
-  dslType: dsl,
-  results: DangerResults,
-  token: string,
-  settings: GitHubRunSettings
-) => {
-  const githubAPI = githubAPIForCommentable(token, settings.repoName!, settings.commentableID)
-  const gh = new GitHub(githubAPI)
-  const platform = perilPlatform(dslType, gh, {})
-  const exec = executorForInstallation(platform, vm2)
-
-  // TODO: Figure what happens here with `git` as being nully,
-  // for one I think it would mean non-sandbox runs cant use inline?
-  await exec.handleResults(results, {} as any)
-}
-
-// This doesn't feel great, but is OK for now
-const getIssueNumber = (json: any): number | null => {
-  if (json.pull_request) {
-    return json.pull_request.number
-  }
-  if (json.issue) {
-    return json.issue.number
-  }
-  return null
-}
-
-export const githubAPIForCommentable = (token: string, repoSlug: string, issueNumber: number | null) => {
-  const githubAPI = new GitHubAPI({ repoSlug, pullRequestID: String(issueNumber) }, token)
-  githubAPI.additionalHeaders = {
-    Accept: "application/vnd.github.machine-man-preview+json",
-  }
-
-  // How can I get this from an API, if we cannot use /me ?
-  // https://api.github.com/repos/PerilTest/PerilPRTester/issues/5/comments
-  // Talked to GH - they know it's an issue.
-  githubAPI.getUserID = () => Promise.resolve(parseInt(PERIL_BOT_USER_ID as string, 10))
-  return githubAPI
 }
 
 export default githubDangerRunner
