+++ source/runner/run.ts
@@ -3,22 +3,21 @@
 // ❯ yarn build; cat source/runner/fixtures/branch-push.json | sed 's/12345/'"$DANGER_GITHUB_API_TOKEN"'/' | env DEBUG="*" node out/runner/index.js
 // ❯
 //
-import logger from "../logger"
 
+import * as exitHook from "async-exit-hook"
+import { GitHub } from "danger/distribution/platforms/GitHub"
 import { contextForDanger } from "danger/distribution/runner/Dangerfile"
 import { jsonToDSL } from "danger/distribution/runner/jsonToDSL"
 import inlineRunner from "danger/distribution/runner/runners/inline"
 
 import { appendPerilContextToDSL, perilObjectForInstallation } from "../danger/append_peril"
 import { dangerRepresentationForPath, dsl } from "../danger/danger_run"
-import { executorForInstallation, InstallationToRun } from "../danger/danger_runner"
+import { executorForInstallation, InstallationToRun, ValidatedPayload } from "../danger/danger_runner"
 import { getPerilPlatformForDSL } from "../danger/peril_platform"
+import { githubAPIForCommentable } from "../github/events/utils/commenting"
 import { getGitHubFileContentsFromLocation } from "../github/lib/github_helpers"
+import logger from "../logger"
 import { PerilRunnerObject } from "./triggerSandboxRun"
-
-import * as exitHook from "async-exit-hook"
-import { GitHub } from "danger/distribution/platforms/GitHub"
-import { githubAPIForCommentable } from "../github/events/github_runner"
 
 let runtimeEnv = {} as any
 
@@ -28,6 +27,7 @@ export const run = async (stdin: string) => {
     return
   }
 
+  // Get STDIN and convert it into a Peril obj
   let input: PerilRunnerObject
   try {
     input = JSON.parse(stdin) as PerilRunnerObject
@@ -39,12 +39,22 @@ export const run = async (stdin: string) => {
   logger.info(`Started run for ${input.path}`)
   const installation = input.installation
 
+  // Validate the JSON
+  if (!input.payload.dsl) {
+    logger.error("Received a payload without github metadata")
+    return
+  }
+
+  // Remove the nulls from the payload as by this point it should be set up
+  const payload = input.payload as ValidatedPayload
   const dslMode = input.dslType === "pr" ? dsl.pr : dsl.import
 
-  if (dslMode === dsl.pr) {
-    await runDangerPR(installation, input)
+  // Run the job
+  if (dslMode === dsl.pr && input.payload.dsl) {
+    await runDangerPR(installation, input, payload)
   } else {
-    await runDangerEvent(installation, input)
+    // Pass the webhook right through
+    await runDangerEvent(installation, input, payload)
   }
 
   return null
@@ -53,66 +63,54 @@ export const run = async (stdin: string) => {
 // There's a lot of redundancy between these, but at least they're somewhat separated in their mental
 // model, used to be much harder to keep track of their differences
 
-const runDangerEvent = async (installation: InstallationToRun, input: PerilRunnerObject) => {
-  // const platform = getPerilPlatformForDSL(dsl.import, null, input.dsl)
-  // const exec = await executorForInstallation(platform, inlineRunner)
-
-  // Create a DSL that is basically just the webhook
-  const token = input.dsl.settings.github.accessToken
-  const context = contextForDanger({ github: input.dsl.github } as any)
+const runDangerEvent = async (installation: InstallationToRun, input: PerilRunnerObject, payload: ValidatedPayload) => {
+  // Pull out the metadata from the JSON to load up the danger process
+  const token = payload.dsl.settings.github.accessToken
+  const context = contextForDanger({ github: payload.dsl.github } as any)
   const peril = perilObjectForInstallation(installation, process.env, input.peril)
 
   // Attach Peril + the octokit API to the DSL
-  // TODO: This probably needs expanding to the util funcs etc
   await appendPerilContextToDSL(installation.iID, token, context, peril)
 
-  const dangerfileLocation = dangerRepresentationForPath(input.path)
-  if (!dangerfileLocation.repoSlug) {
+  const rep = dangerRepresentationForPath(input.path)
+  if (!rep.repoSlug) {
     logger.error(`No repo slug in ${input.path} given for event based run, which is not supported yet`)
     return
   }
 
-  const dangerfileContent = await getGitHubFileContentsFromLocation(
-    token,
-    dangerfileLocation,
-    dangerfileLocation.repoSlug!
-  )
+  const dangerfileContent = await getGitHubFileContentsFromLocation(token, rep, rep.repoSlug!)
 
   runtimeEnv = await inlineRunner.createDangerfileRuntimeEnvironment(context)
-  await inlineRunner.runDangerfileEnvironment(dangerfileLocation.dangerfilePath, dangerfileContent, runtimeEnv)
+  await inlineRunner.runDangerfileEnvironment(rep.dangerfilePath, dangerfileContent, runtimeEnv, payload.webhook)
 }
 
-const runDangerPR = async (installation: InstallationToRun, input: PerilRunnerObject) => {
-  const token = input.dsl.settings.github.accessToken
-
-  if (!input.dsl.github) {
-    logger.info(`Input DSL did not have a github object, ${input}`)
+const runDangerPR = async (installation: InstallationToRun, input: PerilRunnerObject, payload: ValidatedPayload) => {
+  if (!payload.dsl.github) {
+    logger.error("PR payload did not have a github")
     return
   }
 
-  const pr = input.dsl.github.pr
+  const token = payload.dsl.settings.github.accessToken
+  const pr = payload.dsl.github.pr
+
   const perilGHAPI = githubAPIForCommentable(token, pr.base.repo.full_name, pr.number)
   const perilGH = new GitHub(perilGHAPI)
 
-  const platform = getPerilPlatformForDSL(dsl.pr, perilGH, input.dsl)
+  const platform = getPerilPlatformForDSL(dsl.pr, perilGH, payload.dsl)
   const exec = await executorForInstallation(platform, inlineRunner)
 
-  const runtimeDSL = await jsonToDSL(input.dsl)
+  const runtimeDSL = await jsonToDSL(payload.dsl)
   const context = contextForDanger(runtimeDSL)
   const peril = perilObjectForInstallation(installation, process.env, input.peril)
   await appendPerilContextToDSL(installation.iID, token, context, peril)
 
-  const dangerfileLocation = dangerRepresentationForPath(input.path)
+  const rep = dangerRepresentationForPath(input.path)
 
-  const defaultRepoSlug = input.dsl.github.pr.base.repo.full_name
-  const dangerfileContent = await getGitHubFileContentsFromLocation(token, dangerfileLocation, defaultRepoSlug)
+  const defaultRepoSlug = payload.dsl.github.pr.base.repo.full_name
+  const dangerfileContent = await getGitHubFileContentsFromLocation(token, rep, defaultRepoSlug)
 
   runtimeEnv = await inlineRunner.createDangerfileRuntimeEnvironment(context)
-  const results = await inlineRunner.runDangerfileEnvironment(
-    dangerfileLocation.dangerfilePath,
-    dangerfileContent,
-    runtimeEnv
-  )
+  const results = await inlineRunner.runDangerfileEnvironment(rep.dangerfilePath, dangerfileContent, runtimeEnv)
 
   logger.info(
     `f: ${results.fails.length} w: ${results.warnings.length} m: ${results.messages.length} md: ${
