+++ source/scheduler/_tests/_runJob.test.ts
@@ -16,7 +16,7 @@ jest.mock("../../danger/danger_runner", () => ({
 import { runDangerForInstallation } from "../../danger/danger_runner"
 import generateInstallation from "../../testing/installationFactory"
 
-const installation = generateInstallation({ iID: 123 })
+const installation = generateInstallation({ iID: 123, perilSettingsJSONURL: "private/repo@settings.json" })
 const contents = getGitHubFileContents as any
 
 it("runs a dangerfile", async () => {
@@ -24,12 +24,10 @@ it("runs a dangerfile", async () => {
 
   await runJob(installation, "danger/danger-repo@hello.ts")
 
-  expect(runDangerForInstallation).toBeCalledWith("file", "hello.ts", null, dsl.import, installation, {})
+  expect(runDangerForInstallation).toBeCalledWith("file", "hello.ts", null, dsl.import, installation, expect.anything())
 })
 
-jest.mock("../../globals", () => ({ DATABASE_JSON_FILE: "private/repo" }))
-
-it("uses the project settings repo when no repo is passed", async () => {
+it("uses the installation settings repo when no repo is passed", async () => {
   contents.mockImplementationOnce(() => Promise.resolve("file"))
 
   await runJob(installation, "weekly.ts")
